requirements.txt
aiocqhttp
pillow
requests_cache
numpy